+++ docs/source/changelog.rst
@@ -2,6 +2,15 @@
 Releases
 ========
 
+
+Unreleased
+==========
+
+* Seek operations have been restricted to the ``wsgi.input`` of POST and PUT
+  requests. This fixes a bug with Django Admin in Django 1.2 where a view
+  gets the POST arguments even if the request is a GET one.
+
+
 Version 1.0 Final (2010-07-22)
 ==============================